Try fixing broken mouse-yank-primary, again
Since at least 24.3, but maybe earlier, `mouse-yank-primary` is broken
when one selects a region of a buffer in one window and then tries to
paste it with `mouse-yank-primary` into the same buffer in another
window. `mouse-yank-primary` used to handle that correctly, but now it
extends the selection to the point you try pasting to and thus hilarity
ensues or rather cursing.