Revamp frame handling: desktop-save, titles
Emacs 24.4 can save the frame configuration but looses the position
information of frames that are iconified during `desktop-save`.

Reactivate the home-grown `lse-frame:desktop-save` to avoid that
problem.

Totally revamp handling of frame titles: always use `frame-title-format`
to set the frame title, as that allows adapting the frame title to
changes in the selected buffer. Use `title-prefix-suffix` to modify
`frame-title-format`, e.g., for the root-frame and server-window.

Add hooks to `desktop-after-read-hook` and `after-init-hook` to reset
legacy `title-prefix` and frame title parameters.